requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.8
jieba3k==0.35.1
lxml==4.4.1
newspaper3k==0.2.8
nltk==3.4.5
numpy==1.17.2
pandas==0.25.1
Pillow==6.1.0
# pkg-resources==0.0.0
pymongo==3.9.0
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
requests-file==1.4.3
six==1.12.0
soupsieve==1.9.4
tinysegmenter==0.3
tldextract==2.2.1
urllib3==1.25.6
